requirements.txt
torch
torchvision
python-socketio
python-socketio[client]
requests
numpy
Pillow
tqdm
eventlet